Made vehicles predicted
