+++ Content.Server/Resist/EscapeInventorySystem.cs
@@ -11,6 +11,7 @@ using Content.Shared.Storage;
 using Content.Shared.Tag; // Starlight Edit
 using Robust.Server.GameObjects; // Starlight Edit
 using Robust.Shared.Containers;
+using Content.Shared._FarHorizons.Vehicles.Components;
 
 namespace Content.Server.Resist;
 
@@ -42,6 +43,8 @@ public sealed class EscapeInventorySystem : EntitySystem
 
         if (!_containerSystem.TryGetContainingContainer((uid, null, null), out var container) || !_actionBlockerSystem.CanInteract(uid, container.Owner))
             return;
+
+        if (HasComp<RiderComponent>(uid)) return; //FarHorizon
 
         // Make sure there's nothing stopped the removal (like being glued)
         if (!_containerSystem.CanRemove(uid, container))
