+++ Resources/Prototypes/_FarHorizons/Entities/Objects/Specific/Janitorial/janitor.yml
@@ -18,7 +18,7 @@
     - type: ExaminableSolution
       solution: watertank
       exactVolume: true
-    - type: ReagantDraw
+    - type: ReagentDraw
       enabled: false
       solutionContainer: watertank
       whitelistedReagants:
