+++ Resources/Prototypes/_FarHorizons/Entities/Objects/Vehicles/vehicles_gas.yml
@@ -71,7 +71,7 @@
   - type: ExaminableSolution
     solution: fueltank
     exactVolume: true
-  - type: ReagantDraw
+  - type: ReagentDraw
     enabled: false
     solutionContainer: fueltank
     whitelistedReagants: 
@@ -331,7 +331,7 @@
   - type: ExaminableSolution
     solution: fueltank
     exactVolume: true
-  - type: ReagantDraw
+  - type: ReagentDraw
     enabled: false
     solutionContainer: fueltank
     whitelistedReagants:
